% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForestPredict.R
\name{randomForestPredict}
\alias{randomForestPredict}
\title{Predict with the Random Forest model}
\usage{
randomForestPredict(
  model,
  newdata,
  method = NULL,
  bagging.function = mean,
  parallel.plan = NA,
  workers = NULL,
  ...
)
}
\arguments{
\item{model}{A Random Forest, as output from \code{randomForestTrain()}.}

\item{newdata}{The predictors for which to predict \code{newdata}.}

\item{method}{Controls how the prediction is performed. Set this parameter
as one of the following:
\itemize{
\item \code{NULL} (default) performs bagging using the parameter \code{bagging.function}.
\item \code{NA} or \code{"leaves"} returns all the training observations' values falling
on the leaves for each predictor.
\item \code{"random.sample"} draws a non generative random sample from the observations
falling on the leaves.
\item \code{"aposteriori"} uses a posteriori prediction with Moments Matching
Estimation (MME). This is the safest method for \code{fitdistrplus::fitdist()}.
\item The estimation method for \code{fitdistrplus::fitdist()}, i.e. one of the
following: \code{"mle"} for Maximum Likelihood Estimation; \code{"mme"} for
Moments Matching Estimation; \code{"qme"} for Quantile Matching Estimation;
\code{mge} for Maximum Goodness-of-fit Estimation; \code{mse} for
Maximum Spacing Estimation.
}}

\item{bagging.function}{The bagging (aggregation) function if bagging is used,
using \code{method = NULL}. Otherwise ignored.
Use \code{NA} to return the whole set of predictions for all the trees in the forest.}

\item{parallel.plan}{Controls parallel execution, which is handled by package
\code{future.apply}. The default value, \code{NA}, does not use parallel execution. If
set to \code{NULL}, retains the \code{plan()} of the session. Otherwise this parameter
corresponds to \code{plan(strategy = parallel.plan)}. Set it to \code{multisession} for
parallel execution and to \code{sequential} to override plan to sequential.}

\item{workers}{The number of workers. By default uses the maximum available cores.}

\item{...}{Either optional arguments for the bagging function or additional arguments
to \code{fitdistrplus::fitdist()}}
}
\description{
Use the Random Forest model trained with \code{randomForestTrain()} to
predict \code{newdata}.
}
