% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForestSimulate.R
\name{randomForestSimulate}
\alias{randomForestSimulate}
\title{Simulate stochastic series with the Random Forest model predictions.}
\usage{
randomForestSimulate(prediction, n = 1, distr)
}
\arguments{
\item{prediction}{A prediction object, as output by \code{.randomForestPredict()}.
It needs the parameter \code{simplify.estimation} to be set as \code{TRUE}
\code{randomForestPredict()}.}

\item{n}{How many series to simulate?}

\item{distr}{The probability distribution. Leave empty or set to \code{NULL} for
automatically selecting the distribution. Currently supported distributions are the
bernoulli distribution, \code{"bernoulli"}; the normal distribution, \code{"norm"};
the gamma distribution, \code{"gamma"}; and the exponential distribution, \code{"exp"}.}
}
\description{
Use the predictions output by \code{randomForestPredict()} to simulate
\code{n} stochastic series.
}
