% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForestTrain.R
\name{randomForestTrain}
\alias{randomForestTrain}
\title{Train Random Forest model}
\usage{
randomForestTrain(
  x,
  y = NULL,
  ntree = 100,
  mtry = if (!is.null(y) && !is.factor(y)) max(floor(ncol(x)/3), 1) else
    floor(sqrt(ncol(x))),
  minsplit = if (!is.null(y) && !is.factor(y)) 5 else 1,
  minbucket = minsplit/3,
  maxdepth = 30,
  method,
  replace = TRUE,
  sampsize = if (replace) nrow(x) else ceiling(0.632 * nrow(x)),
  parallel.plan,
  workers,
  weights,
  parms,
  remove.leaf.info = FALSE
)
}
\arguments{
\item{x}{The predictors, in matrix form.}

\item{y}{The predictands, a vector.}

\item{ntree}{Number of trees.}

\item{mtry}{Number of predictors randomly used as candidate split.
Default is \code{sqrt(ncol(x))} for classification and \code{ncol(x)/3} for
regression}

\item{minsplit}{Minimum number of elements to attempt a split.}

\item{minbucket}{Minimum terminal leaf size, i.e. the minimum number of observations
that must be present on a terminal leaf.}

\item{maxdepth}{Maximum depth of the trees.}

\item{method}{The split function. Currently supported split functions are:
\code{"anova"} (default) for root mean squared error; \code{"poisson"} for poisson
distributed \code{y} (expect two columns); \code{"class"} for classification of
factors (default if \code{y} is factor); \code{"exp"} for survival objects;
\code{"gammaLLMME"}, \code{"gammaLLmean"}, \code{"gammaLLBC3"} for the 2-parameter
gamma distribution log-likelihood using moments-matching-estimation, the mean, and
the BC3 estimators, respectively; \code{"bernoulliGammaLLMME"} for the 3-parameter
bernoulli-gamma log likelihood; \code{"gammaDeviation"} for the deviation of the
2-parameter gamma distribution; \code{"bernoulliLL"} for the bernoulli
log-likelihood (expects class \code{numeric} with \code{0,1}).}

\item{replace}{Either \code{TRUE} or \code{FALSE}. Should samples of \code{x} be
drawn with replacement?}

\item{sampsize}{Size of the samples of \code{x}}

\item{parallel.plan}{Controls parallel execution, which is handled by and requires the
package \code{future.apply}. If this package is not installed parallel execution will
 not be used. If left missing or set to \code{NULL} (default), the function uses the
current \code{future::plan()} of the session. Set this parameter to \code{"auto"} for
automatic parallelization, which avoids dealing with futures outside the function.
In any other case this parameter corresponds to
\code{futute::plan(strategy = parallel.plan)}, check their corresponding help pages.
\code{NA} avoids the use of \code{future.apply::future_lapply()}.}

\item{workers}{The number of workers. By default uses the maximum available cores.}

\item{weights}{\code{weights} as pased to \code{rpart::rpart()}.}

\item{parms}{\code{parms} as pased to \code{rpart::rpart()}.}

\item{remove.leaf.info}{Reduces the memory usage of the trees. This should not be
set to \code{TRUE} unless there are memory issues. NOTE: Setting this to \code{TRUE}
makes a posteriori estimation not available.}
}
\description{
Train a random forest for predictors \code{x} and predictands \code{y}.
For each tree a bootstrap sample is drawn from \code{x}. By default these samples are
drawn with replacement (\code{replace = TRUE}), and have size \code{.632*nrow(x)}.
}
