% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForestPredict.R
\name{randomForestPredict}
\alias{randomForestPredict}
\title{Predict with the Random Forest model}
\usage{
randomForestPredict(
  model,
  newdata,
  method = NULL,
  bagging.function = mean,
  distr = NULL,
  simplify.estimation = TRUE,
  ...
)
}
\arguments{
\item{model}{A Random Forest, as output from \code{randomForestTrain()}.}

\item{newdata}{The predictors for which to predict \code{newdata}.}

\item{method}{Controls how the prediction is performed. Set this parameter
as one of the following:
\itemize{
\item \code{NULL} (default) performs bagging using the parameter \code{bagging.function}.
\item \code{NA} or \code{"leaves"} returns all the training observations' values falling
on the leaves for each predictor.
\item \code{"random.sample"} draws a non generative random sample from the observations
falling on the leaves.
\item \code{"aposteriori"} uses a posteriori prediction with Moments Matching
Estimation (MME). This is the safest method for \code{fitdistrplus::fitdist()}.
\item The estimation method for \code{fitdistrplus::fitdist()}, i.e. one of the
following: \code{"mle"} for Maximum Likelihood Estimation; \code{"mme"} for
Moments Matching Estimation; \code{"qme"} for Quantile Matching Estimation;
\code{mge} for Maximum Goodness-of-fit Estimation; \code{mse} for
Maximum Spacing Estimation.
}}

\item{bagging.function}{The bagging (aggregation) function if bagging is used,
using \code{method = NULL}. Otherwise ignored.
Use \code{NA} to return the whole set of predictions for all the trees in the forest.}

\item{distr}{Only used if using aposteriori estimation
(\code{method = "aposteriori", "mle", ...}. The probability distribution to be estimated,
parameter passed to \code{fitdistrplus::fitdist()}. Will be automatically infered if set
to \code{NULL}.}

\item{simplify.estimation}{Only used if using aposteriori estimation. Set to \code{FALSE}
to return the whole output from \code{fitdistrplus::fitdist()}.}

\item{...}{Either optional arguments for the bagging function or additional arguments
to \code{fitdistrplus::fitdist()}}
}
\description{
Use the Random Forest model trained with \code{randomForestTrain()} to
predict \code{newdata}.
}
