% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForestTrain.R
\name{randomForestTrain}
\alias{randomForestTrain}
\title{Train Random Forest model}
\usage{
randomForestTrain(
  x,
  y = NULL,
  ntree = 100,
  mtry = if (!is.null(y) && !is.factor(y)) max(floor(ncol(x)/3), 1) else
    floor(sqrt(ncol(x))),
  minsplit = if (!is.null(y) && !is.factor(y)) 5 else 1,
  minbucket = minsplit/3,
  maxdepth = 30,
  method,
  replace = TRUE,
  sampsize = if (replace) nrow(x) else ceiling(0.632 * nrow(x)),
  parallel.plan,
  workers,
  weights,
  parms,
  remove.leaf.info = FALSE
)
}
\arguments{
\item{x}{The predictors, in matrix form.}

\item{y}{The predictands, a vector.}

\item{ntree}{Number of trees.}

\item{mtry}{Number of predictors randomly used as candidate split.
Default is \code{sqrt(ncol(x))} for classification and \code{ncol(x)/3} for
regression}

\item{minsplit}{Minimum number of elements to attempt a split.}

\item{minbucket}{Minimum terminal leaf size, i.e. the minimum number of observations
that must be present on a terminal leaf.}

\item{maxdepth}{Maximum depth of the trees.}

\item{method}{The split function.}

\item{replace}{Either \code{TRUE} or \code{FALSE}. Should samples of \code{x} be
drawn with replacement?}

\item{sampsize}{Size of the samples of \code{x}}

\item{parallel.plan}{Controls parallel execution, which is handled by package
\code{future.apply}. The default value, \code{NA}, does not use parallel execution. If
set to \code{NULL}, retains the \code{plan()} of the session. Otherwise this parameter
corresponds to \code{plan(strategy = parallel.plan)}. Set it to \code{multisession} for
parallel execution and to \code{sequential} to override plan to sequential.}

\item{workers}{The number of workers. By default uses the maximum available cores.}

\item{weights}{\code{weights} as pased to \code{rpart::rpart()}.}

\item{parms}{\code{parms} as pased to \code{rpart::rpart()}.}

\item{remove.leaf.info}{Reduces the memory usage of the trees. This should not be
set to \code{TRUE} unless there are memory issues. NOTE: Setting this to \code{TRUE}
makes a posteriori estimation not available.}
}
\description{
Train a random forest for predictors \code{x} and predictands \code{y}.
For each tree a bootstrap sample is drawn from \code{x}. By default these samples are
drawn with replacement (\code{replace = TRUE}), and have size \code{.632*nrow(x)}.
}
